+++ spatial/r3/vector.go
@@ -55,9 +55,6 @@ func (p Vec) Cross(q Vec) Vec {
 
 // Rotate returns a new vector, rotated by alpha around the provided axis.
 func (p Vec) Rotate(alpha float64, axis Vec) Vec {
-	if alpha == 0 {
-		return p
-	}
 	return NewRotation(alpha, axis).Rotate(p)
 }
 
@@ -94,4 +91,47 @@ func Cos(p, q Vec) float64 {
 // Box is a 3D bounding box.
 type Box struct {
 	Min, Max Vec
+}
+
+// TODO:
+//  - create rotations from Euler angles (NewRotationFromEuler?)
+//  - create rotations from rotation matrices (NewRotationFromMatrix?)
+//  - return the equivalent Euler angles from a Rotation
+//  - return the equivalent rotation matrix from a Rotation
+
+// Rotation describes a rotation in space.
+type Rotation quat.Number
+
+// NewRotation creates a rotation by alpha, around axis.
+func NewRotation(alpha float64, axis Vec) Rotation {
+	if alpha == 0 {
+		return Rotation(quat.Number{
+			Real: 1, Imag: 1, Jmag: 1, Kmag: 1,
+		})
+	}
+	q := raise(axis)
+	sin, cos := math.Sincos(0.5 * alpha)
+	q = quat.Scale(sin/quat.Abs(q), q)
+	q.Real += cos
+	if len := quat.Abs(q); len != 1 {
+		q = quat.Scale(1/len, q)
+	}
+
+	return Rotation(q)
+}
+
+// Rotate returns the rotated vector according to the definition of rot.
+func (r Rotation) Rotate(p Vec) Vec {
+	if r.isIdentity() {
+		return p
+	}
+	qq := quat.Number(r)
+	pp := quat.Mul(quat.Mul(qq, raise(p)), quat.Conj(qq))
+	return Vec{X: pp.Imag, Y: pp.Jmag, Z: pp.Kmag}
+}
+
+func (r Rotation) isIdentity() bool {
+	return quat.Number(r) == quat.Number{
+		Real: 1, Imag: 1, Jmag: 1, Kmag: 1,
+	}
 }
